requirements.txt
absl-py==1.3.0
accelerate==0.15.0
aiohttp==3.8.3
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.1.0
beartype==0.11.0
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cmake==3.26.1
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
datasets==2.7.1
dill==0.3.6
docker-pycreds==0.4.0
einops==0.6.0
einops-exts==0.0.3
ema-pytorch==0.1.2
et-xmlfile==1.1.0
filelock==3.8.2
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.15.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
h5py==3.8.0
huggingface-hub==0.11.1
idna==3.4
imageio==2.22.4
imagen-pytorch==1.18.5
importlib-metadata==5.1.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
kornia==0.6.8
lightning-utilities==0.4.2
lit==16.0.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
mpmath==1.3.0
multidict==6.0.3
multiprocess==0.70.14
networkx==2.8.8
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.7.0.72
openpyxl==3.0.10
packaging==22.0
pandas==1.5.2
pathtools==0.1.2
Pillow==9.3.0
protobuf==3.20.1
psutil==5.9.4
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-fid==0.3.0
pytorch-lightning==1.8.4.post0
pytorch-warmup==0.1.1
pytz==2022.6
PyWavelets==1.4.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
scikit-image==0.19.3
scipy==1.9.3
sentencepiece==0.1.97
sentry-sdk==1.18.0
setproctitle==1.3.2
six==1.16.0
slideio==2.0.8
smmap==5.0.0
sympy==1.11.1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tifffile==2022.10.10
tokenizers==0.13.2
torch==2.0.0+cu118
torchaudio==2.0.1+cu118
torchmetrics==0.11.0
torchvision==0.15.1+cu118
tqdm==4.64.1
transformers==4.25.1
triton==2.0.0
typing-extensions==4.4.0
urllib3==1.26.13
wandb==0.14.0
Werkzeug==2.2.2
xxhash==3.1.0
yarl==1.8.2
zipp==3.11.0
